+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.2] - 2025-06-28
 
-- Support for `--help` and `--debug` flags to studio-mcp itself.
+- Support for `--help` and `--debug` flags to studio itself.
 - Improved handling of templated command arguments.
 
 ## [0.0.1] - 2025-06-28